requirements.txt
# Python dependencies for AI-Powered DevOps/MLOps Pipeline Optimizer

# Core dependencies
pyyaml>=6.0
python-dateutil>=2.8.2

# Async support
aiohttp>=3.8.0
asyncio>=3.4.3

# Data processing
pandas>=1.5.0
numpy>=1.24.0

# AI/ML (optional - for advanced features)
# openai>=1.0.0  # Uncomment if using OpenAI API
# anthropic>=0.5.0  # Uncomment if using Claude API

# API clients for pipeline integrations
requests>=2.31.0

# Testing
pytest>=7.4.0
pytest-asyncio>=0.21.0
pytest-cov>=4.1.0

# Logging and monitoring
python-json-logger>=2.0.7

# Optional: Web dashboard dependencies (uncomment if needed)
# flask>=3.0.0
# dash>=2.14.0
# plotly>=5.18.0

# Optional: Notification integrations (uncomment if needed)
# slack-sdk>=3.23.0
# python-telegram-bot>=20.6

# Optional: Cloud platform SDKs (uncomment as needed)
# boto3>=1.28.0  # AWS
# google-cloud-storage>=2.10.0  # GCP
# azure-storage-blob>=12.19.0  # Azure

# Optional: Pipeline platform clients (uncomment as needed)
# python-jenkins>=1.8.2
# python-gitlab>=4.2.0
# mlflow>=2.9.0
# kubernetes>=28.1.0
